efk-stack/python-app/requirements.txt
fluent-logger

